misc refactor/cleanup
- remove v1 accessions generated types
- move inventory summary fetch into summary component
- block the flashing empty state in inventory table
